rename : 폴더 구조 정리 - Controller, Service, Docs가 2개 이상이면 폴더로 정리
