+++ src/place/place.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { PlaceController } from './place.controller';
-import { PlaceService } from './place.service';
+import { PlaceController } from './controllers/place.controller';
+import { PlaceService } from './services/place.service';
 import { PlaceRepository } from 'src/place/repositories/place.repository';
 import { PlaceImageRepository } from 'src/place/repositories/place-image.repository';
 import { TagModule } from 'src/tag/tag.module';
-import { S3Service } from 'src/place/s3.service';
+import { S3Service } from 'src/place/services/s3.service';
 import { PlaceTagRepository } from './repositories/place-tag.repository';
 import { PlaceDetailRepository } from './repositories/place-detail.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
@@ -15,9 +15,9 @@ import { PlaceTag } from './entities/place-tag.entity';
 import { UserModule } from 'src/user/user.module';
 import { CollectionModule } from 'src/collection/collection.module';
 import { KakaoCategoryMapping } from './entities/kakao-category-mapping.entity';
-import { KakaoCategoryMappingService } from './kakao-category-mapping.service';
 import { KakaoCategoryMappingRepository } from './repositories/kakao-category-mapping.repository';
-import { KakaoCategoryMappingController } from './kakao-category-mapping.controller';
+import { KakaoCategoryMappingController } from './controllers/kakao-category-mapping.controller';
+import { KakaoCategoryMappingService } from './services/kakao-category-mapping.service';
 
 @Module({
   controllers: [PlaceController, KakaoCategoryMappingController],
