+++ src/place/services/s3.service.ts
